Rename extension traits according to rfc 445
